implemented menu and its changes to edit/save/delete entries
Added menu to JournalActivity with logic to show edit/delete when an entry is longclicked. Choosing edit converts the textView to editText, along with changing the menu options to allow the user to make changes. Both menu variants offer a cancel "X" to allow the user to leave the modus.

Also implemented dialogs to warn for unsaved changed or before deleting an entry.